Move `quorum_certificate` classes to `hotstuff.hpp`, and add `hotstuff.cpp`
