+++ libraries/chain/include/eosio/chain/hotstuff/hotstuff.hpp
@@ -116,6 +116,122 @@ namespace eosio::chain {
       }
    };
 
+   using bls_public_key  = fc::crypto::blslib::bls_public_key;
+   using bls_signature   = fc::crypto::blslib::bls_signature;
+   using bls_private_key = fc::crypto::blslib::bls_private_key;
+
+   // -------------------- pending_quorum_certificate -------------------------------------------------
+   class pending_quorum_certificate {
+   public:
+      enum class state_t {
+         unrestricted,  // No quorum reached yet, still possible to achieve any state.
+         restricted,    // Enough `weak` votes received to know it is impossible to reach the `strong` state.
+         weak_achieved, // Enough `weak` + `strong` votes for a valid `weak` QC, still possible to reach the `strong` state.
+         weak_final,    // Enough `weak` + `strong` votes for a valid `weak` QC, `strong` not possible anymore.
+         strong         // Enough `strong` votes to have a valid `strong` QC
+      };
+
+      struct votes_t {
+         hs_bitset     _bitset;
+         bls_signature _sig;
+
+         void resize(size_t num_finalizers) { _bitset.resize(num_finalizers); }
+         size_t count() const { return _bitset.count(); }
+
+         bool add_vote(const std::vector<uint8_t>& proposal_digest, size_t index, const bls_public_key& pubkey,
+                       const bls_signature& new_sig);
+
+         void reset(size_t num_finalizers);
+      };
+
+      pending_quorum_certificate() = default;
+
+      explicit pending_quorum_certificate(size_t num_finalizers, size_t quorum);
+
+      explicit pending_quorum_certificate(const fc::sha256& proposal_id,
+                                          const digest_type& proposal_digest,
+                                          size_t num_finalizers,
+                                          size_t quorum);
+
+      size_t num_weak()   const { return _weak_votes.count(); }
+      size_t num_strong() const { return _strong_votes.count(); }
+
+      bool   is_quorum_met() const;
+
+      void reset(const fc::sha256& proposal_id, const digest_type& proposal_digest, size_t num_finalizers, size_t quorum);
+
+      bool add_strong_vote(const std::vector<uint8_t>& proposal_digest,
+                           size_t index,
+                           const bls_public_key& pubkey,
+                           const bls_signature& sig);
+
+      bool add_weak_vote(const std::vector<uint8_t>& proposal_digest,
+                         size_t index,
+                         const bls_public_key& pubkey,
+                         const bls_signature& sig);
+
+      bool add_vote(bool strong,
+                    const std::vector<uint8_t>& proposal_digest,
+                    size_t index,
+                    const bls_public_key& pubkey,
+                    const bls_signature& sig);
+
+      // ================== begin compatibility functions =======================
+      // these are present just to make the tests still work. will be removed.
+      // these assume *only* strong votes.
+      quorum_certificate_message to_msg() const;
+      const fc::sha256&          get_proposal_id() const { return _proposal_id; }
+      std::string                get_votes_string() const;
+      // ================== end compatibility functions =======================
+
+      friend struct fc::reflector<pending_quorum_certificate>;
+      fc::sha256           _proposal_id;     // only used in to_msg(). Remove eventually
+      std::vector<uint8_t> _proposal_digest;
+      state_t              _state { state_t::unrestricted };
+      size_t               _num_finalizers {0};
+      size_t               _quorum {0};
+      votes_t              _weak_votes;
+      votes_t              _strong_votes;
+   };
+
+   // -------------------- valid_quorum_certificate -------------------------------------------------
+   class valid_quorum_certificate {
+   public:
+      valid_quorum_certificate(const pending_quorum_certificate& qc);
+
+      valid_quorum_certificate(const fc::sha256& proposal_id,
+                               const std::vector<uint8_t>& proposal_digest,
+                               const std::vector<uint32_t>& strong_votes, //bitset encoding, following canonical order
+                               const std::vector<uint32_t>& weak_votes,   //bitset encoding, following canonical order
+                               const bls_signature& sig);
+
+      valid_quorum_certificate() = default;
+      valid_quorum_certificate(const valid_quorum_certificate&) = default;
+
+      bool is_weak()   const { return !!_weak_votes; }
+      bool is_strong() const { return !_weak_votes; }
+
+      // ================== begin compatibility functions =======================
+      // these are present just to make the tests still work. will be removed.
+      // these assume *only* strong votes.
+      quorum_certificate_message to_msg() const;
+      const fc::sha256&          get_proposal_id() const { return _proposal_id; }
+      // ================== end compatibility functions =======================
+
+      friend struct fc::reflector<valid_quorum_certificate>;
+      fc::sha256               _proposal_id;     // [todo] remove
+      std::vector<uint8_t>     _proposal_digest; // [todo] remove
+      std::optional<hs_bitset> _strong_votes;
+      std::optional<hs_bitset> _weak_votes;
+      bls_signature            _sig;
+   };
+
+   // -------------------- quorum_certificate -------------------------------------------------------
+   struct quorum_certificate {
+      uint32_t block_height;
+      valid_quorum_certificate qc;
+   };
+
 } //eosio::chain
